fixup: Simplified swap_hdr_len handling
... by moving casts from callers to SwapMetaUnpacker itself.

Also changed staleSwapHeaderSize and related types from uint64_t to
size_t. Technically, uint64_t may be the best long-term choice because
we may end up supporting very long Store meta field lists, with large
fields (e.g., up to storing the HTTP response body as a field). However,
all the current de-serializing code assumes that store metadata fits a
single buffer, making size_t a better choice (for now) because it would
avoid additional overflow checks.

+++ src/fs/rock/RockHeaderUpdater.h
@@ -68,7 +68,7 @@ private:
     SBuf exchangeBuffer; ///< bytes read but not yet discarded or written
     uint64_t bytesRead; ///< total entry bytes read from Store so far
 
-    uint64_t staleSwapHeaderSize; ///< stored size of the stale entry metadata
+    size_t staleSwapHeaderSize; ///< stored size of the stale entry metadata
 
     SlotId staleSplicingPointNext; ///< non-updatable old HTTP body suffix start
 };
